+++ lib/eve_online/esi/character_ship.rb
@@ -1,0 +1,36 @@
+# frozen_string_literal: true
+
+require 'forwardable'
+
+module EveOnline
+  module ESI
+    class CharacterShip < Base
+      extend Forwardable
+
+      API_ENDPOINT = 'https://esi.evetech.net/v1/characters/%<character_id>s/ship/?datasource=%<datasource>s'
+
+      attr_reader :character_id
+
+      def initialize(options)
+        super
+
+        @character_id = options.fetch(:character_id)
+      end
+
+      def_delegators :model, :as_json, :ship_item_id, :ship_name, :ship_type_id
+
+      def model
+        Models::CharacterShip.new(response)
+      end
+      memoize :model
+
+      def scope
+        'esi-location.read_ship_type.v1'
+      end
+
+      def url
+        format(API_ENDPOINT, character_id: character_id, datasource: datasource)
+      end
+    end
+  end
+end